+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+
+- chore(databases): simplify fork error description
+
 ## [0.11.0] - 2026-09-18
 
 ### Removed